feature: create sale order and remove product quantity

+++ Capstone.Application/Services/Products/IProductsService.cs
@@ -36,8 +36,8 @@ public interface IProductsService
         List<ProductQuantityDto> productQuantities,
         string createdBy,
         IFormFile? image,
-        decimal salePrice,
-        decimal importPrice
+        double salePrice,
+        double importPrice
     );
 
     Task<Result<ProductDto>> OwnerPatchProduct(
@@ -49,8 +49,8 @@ public interface IProductsService
         string? pattern,
         string? sizeType,
         List<ProductQuantityDto>? quantities,
-        decimal? salePrice,
-        decimal? importPrice
+        double? salePrice,
+        double? importPrice
     );
 
     Task<Result<PaginatedResult<ProductDto>>> FetchAllProducts(int currentPage, int pageSize, string? category = null, string? search = null);
@@ -63,8 +63,8 @@ public interface IProductsService
         string? pattern,
         string? sizeType,
         List<ProductQuantityDto>? newQuantities,
-        decimal? salePrice,
-        decimal? importPrice
+        double? salePrice,
+        double? importPrice
     );
 
     Task<Result<ProductWithQuantityChangesDto>> EmployeeUpdateProductInProductsOrder(
